+++ enterprise/frontend/src/metabase-enterprise/custom_viz/index.ts
@@ -8,9 +8,9 @@ import {
 import { hasPremiumFeature } from "metabase-enterprise/settings";
 import { isCustomVizDisplay } from "metabase-types/guards/visualization";
 
-import { AddCustomVizPage } from "./components/AddCustomVizPage";
 import { CustomVizDevPage } from "./components/CustomVizDevPage";
 import { CustomVizPage } from "./components/CustomVizPage";
+import { ManageCustomVizPage } from "./components/ManageCustomVizPage";
 import {
   loadCustomVizPlugin,
   useAutoLoadCustomVizPlugin,
@@ -53,8 +53,8 @@ function registerCustomVizPlugin(
 export function initializePlugin() {
   if (hasPremiumFeature("custom-viz")) {
     Object.assign(PLUGIN_CUSTOM_VIZ, {
+      ManageCustomVizPage,
       CustomVizPage,
-      AddCustomVizPage,
       CustomVizDevPage,
       useAutoLoadCustomVizPlugin,
       useCustomVizPlugins,
